Create forked.java
Lunchbox is done with playing chess! His queen and king just got forked again!

In chess, a fork is when a knight attacks two pieces of higher value, commonly the king and the queen. Lunchbox knows that knights can be tricky, and in the version of chess that he is playing, knights are even trickier: instead of moving 1
 tile in one direction and 2
 tiles in the other, knights in Lunchbox's modified game move 𝑎
 tiles in one direction and 𝑏
 tiles in the other.

Lunchbox is playing chess on an infinite chessboard which contains all cells (𝑥,𝑦)
 where 𝑥
 and 𝑦
 are (possibly negative) integers. Lunchbox's king and queen are placed on cells (𝑥𝐾,𝑦𝐾)
 and (𝑥𝑄,𝑦𝑄)
 respectively. Find the number of positions such that if a knight was placed on that cell, it would attack both the king and queen.

Input
Each test contains multiple test cases. The first line contains an integer 𝑡
 (1≤𝑡≤1000
) — the number of test cases. The description of the test cases follows.

The first line of each test case contains two integers 𝑎
 and 𝑏
 (1≤𝑎,𝑏≤108
) — describing the possible moves of the knight.

The second line of each test case contains two integers 𝑥𝐾
 and 𝑦𝐾
 (0≤𝑥𝐾,𝑦𝐾≤108
) — the position of Lunchbox's king.

The third line in a test case contains 𝑥𝑄
 and 𝑦𝑄
 (0≤𝑥𝑄,𝑦𝑄≤108
) — the position of Lunchbox's queen.

It is guaranteed that Lunchbox's queen and king will occupy different cells. That is, (𝑥𝐾,𝑦𝐾)≠(𝑥𝑄,𝑦𝑄)
.

Output
For each test case, output the number of positions on an infinite chessboard such that a knight can attack both the king and the queen.