Create chemistry.java
You are given a string 𝑠
 of length 𝑛
, consisting of lowercase Latin letters, and an integer 𝑘
.

You need to check if it is possible to remove exactly 𝑘
 characters from the string 𝑠
 in such a way that the remaining characters can be rearranged to form a palindrome. Note that you can reorder the remaining characters in any way.

A palindrome is a string that reads the same forwards and backwards. For example, the strings "z", "aaa", "aba", "abccba" are palindromes, while the strings "codeforces", "reality", "ab" are not.

Input
Each test consists of multiple test cases. The first line contains a single integer 𝑡
 (1≤𝑡≤104
) — the number of the test cases. This is followed by their description.

The first line of each test case contains two integers 𝑛
 and 𝑘
 (0≤𝑘<𝑛≤105
) — the length of the string 𝑠
 and the number of characters to be deleted.

The second line of each test case contains a string 𝑠
 of length 𝑛
, consisting of lowercase Latin letters.

It is guaranteed that the sum of 𝑛
 over all test cases does not exceed 2⋅105
.

Output
For each test case, output "YES" if it is possible to remove exactly 𝑘
 characters from the string 𝑠
 in such a way that the remaining characters can be rearranged to form a palindrome, and "NO" otherwise.

You can output the answer in any case (uppercase or lowercase). For example, the strings "yEs", "yes", "Yes", and "YES" will be recognized as positive answers.