Create vasilijeInCacak.java
Aca and Milovan, two fellow competitive programmers, decided to give Vasilije a problem to test his skills.

Vasilije is given three positive integers: 𝑛
, 𝑘
, and 𝑥
, and he has to determine if he can choose 𝑘
 distinct integers between 1
 and 𝑛
, such that their sum is equal to 𝑥
.

Since Vasilije is now in the weirdest city in Serbia where Aca and Milovan live, Cacak, the problem seems weird to him. So he needs your help with this problem.

Input
The first line contains a single integer 𝑡
 (1≤𝑡≤104
) — the number of test cases.

The only line of each test case contains three integers 𝑛
, 𝑘
 and 𝑥
 (1≤𝑛≤2⋅105
, 1≤𝑘≤𝑛
, 1≤𝑥≤4⋅1010
) — the maximum element he can choose, the number of elements he can choose and the sum he has to reach.

Note that the sum of 𝑛
 over all test cases may exceed 2⋅105
.

Output
For each test case output one line: "YES", if it is possible to choose 𝑘
 distinct integers between 1
 and 𝑛
, such that their sum is equal to 𝑥
, and "NO", if it isn't.

You can output the answer in any case (for example, the strings "yEs", "yes", "Yes", and "YES" will be recognized as a positive answer).